The first commit includes the robot's urdf, the model trained by stage-based reward, and the actual hitting trajectory.
The committed models include stage 1 training model, stage 1+2 training model, stage 1+2+3 training model and only stage 3 training model.

In addition, the speed of the ping-pong ball during training is about 3~6m/s, and many of the actual ball speeds exceed 8m/s, which leads to the inferior performance of the model in the reality, so I may re-train the model in the future.